% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC_helper.R
\name{align}
\alias{align}
\title{Cluster Aligner}
\usage{
align(refL, L, type = "vec")
}
\arguments{
\item{refL}{vector or matrix, reference label.}

\item{L}{vector or matrix, label to be aligned.}

\item{type}{what type of input format: "\code{vec}" or "\code{mat}", default as 'vec'.}
}
\value{
\code{L}, aligned against \code{refL} with maximal concordance.
}
\description{
\code{align} aligns two cluster assignments by a maximally concordant relabeling, implemented
  based on Cpp function \code{\link{clustalign}}
}
\details{
This aligner can take both \emph{vector} and \emph{matrix} clustering assignments, and aligns
  the second assignment against the first assignment. It reshuffles the second assignment labels
  such that two assignemnts are maximally matched. This procedure facilitates a direct and
  fair comparision between two clustering result.
}
\examples{
ref <- c(1:5)
L <- c(5:1)

align(ref, L, type = 'vec')

}
