% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEGplot.R
\name{EEGplot}
\alias{EEGplot}
\title{128-EEGNet channel plots}
\usage{
EEGplot(clust, color = NULL)
}
\arguments{
\item{clust}{vector, cluster labels for 124 channels.}

\item{color}{vector, colors for K clusters in the plot.}
}
\description{
\code{EEGplot} plots the clustering results of 124 channels on a 2D projected scalp plot.
}
\examples{
\dontrun{
# The channel plot:

EEGplot(rep(1,124))

}

}
