% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EigLap}
\alias{EigLap}
\title{Eigen Laplacian transformation}
\usage{
EigLap(data, D, normal)
}
\arguments{
\item{data, }{3 dimensional array of similarity matrices}

\item{D, }{integer of the dimensionality transforming into}

\item{normal, }{logical of whether to perform normalization on transformed data.}
}
\value{
3d array admissible to \code{\link{MIC}}
}
\description{
\code{EigLap} operates on symmatrical similarity matrices seeking a \code{D}
  dimensional representation using its first \code{D} eigen-vectors of its
  graph Laplacians.
}
\references{
Andew Ng, Michael Jordan and Yair Weiss "\emph{On Spectral Clustering:
  Analysis and an Algorithm}"
}
\seealso{
\code{\link{MIC_prep}} for its usage.
}
