% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC.R
\name{MIC}
\alias{MIC}
\title{Multilevel Integrative Clustering}
\usage{
MIC(data, K, nit, thin = 1, drop = T)
}
\arguments{
\item{data}{list of 3d arrays, each of which has the size of d(measurement dims)*p(number of units)*r(number of repetitions).
MIC requires \code{p} the same across arrays. For an admissible format, see \code{\link{MIC_prep}}.}

\item{K}{integer, number of clusters.}

\item{nit}{number of iterations for MCMC sampling, with 1/5 used for burn-in.}

\item{thin}{thining, default at 1 (no thinning).}

\item{drop}{logical, to remove posterior samples (default TRUE).}
}
\value{
A list of objects with the following components:
  \item{\code{clustS}}{group level clustering result}
  \item{\code{clustC}}{individual level clustering result}
  \item{\code{clustL}}{repetition level clustering result}
  \item{\code{mixpr}}{cluster proportions with its posterior summaries}
  \item{\code{alpha}}{S-C level adherence and its posterior summaries}
  \item{\code{beta}}{C-L level adherence and its posterior summaries}
  \item{\code{ICs}}{model assessment measures: BIC and adjusted adherence}
}
\description{
\code{MIC} implements integrative clustering on highly structured data. The admissible structure contains repeated measurements
  of the same set of units collected from multiple individuals. Clustering is performed integratively therefore interpretable
  on all three levels (group/individual/repetition).
}
\examples{

\dontrun{
  # Time Series simulation:
  ts_sim <- MIC_sim(alpha = 0.9, nsub = 3, segs = 10, fs = 100)


  # Data preparation:
  list_data <- lapply(ts_sim$Data, function(x)
       MIC_prep(X = x, d = 4, par.spec = c(50, 50, 100), par.win = c(3, 1)))


  # MIC: (Running time: 11s, 30X faster than version 1)
  output <- MIC(data = list_data, K = 4, nit = 10000)



  # Clustering accuracy: group level (to be updated)
  sum(align(ts_sim$C, output$clustS, type = 'vec') == ts_sim$C) / 40
  ## [1] 1


  # Clustering accuracy: individual level (to be updated)
  true_c <- lapply(1:3,function(i) ts_sim$Ci[,i]);
  est_c <- lapply(1:3,function(i) output$clustC[i,]);
  mapply(function(x,y) sum(align(x,y,'vec')==x)/40, true_c, est_c);
  ## [1] 1 1 1
}
}
\references{
Qian Li, Damla Senturk, Catherine A. Sugar, Shanali Jeste, Charlotte DiStefano, Joel Frohlich, Donatello Telesca
  "\emph{Inferring Brain Signals Synchronicity from a Sample of EEG Readings}".
}
\seealso{
\code{\link{MIC_prep}} for data preparation, \code{\link{MIC_sim}} for MIC on simulated time series.
}
