% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MIC_mcmc}
\alias{MIC_mcmc}
\title{MIC sampling function in C++}
\usage{
MIC_mcmc(data, K, run, thin)
}
\arguments{
\item{data, }{R list of 3D array, see \code{\link{MIC}}}

\item{K, }{integer as number of clusters}

\item{run, }{integer as number of MCMC samples}

\item{thin, }{integer as thinning every few iterations}
}
\description{
\code{MIC_mcmc} performs Gibbs sampling in C++ as an essential part of
  \code{\link{MIC}}
}
