% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC_sim.R
\name{MIC_sim}
\alias{MIC_sim}
\title{Time Series Simulation for MIC}
\usage{
MIC_sim(alpha, nsub, segs, fs, Ct = rep(c(1, 2, 3, 4), rep(10, 4)),
  bad_sub = 0, scheme = 0, iSNR = 0)
}
\arguments{
\item{alpha}{numeric in [0, 1], group-subject adherence}

\item{nsub}{integer, number of subjects}

\item{segs}{integer, number of segments observed}

\item{fs}{integer, sampling frequency}

\item{Ct}{vector of integers, true group labels, defuault at 4 groups each with 10 subjects}

\item{bad_sub}{integer, number of outlying subjects indexed from 1 to \code{bad_sub}}

\item{scheme}{stationarity: 0(default) = stationary; 1 = piecewise stationary}

\item{iSNR}{scalar, inverse of Signal-to-Noise Ratio, default at 0 (no noise)}
}
\value{
A list of objects with the following components:
  \item{\code{C}}{True group label}
  \item{\code{Ci}}{True individual label}
  \item{\code{Data}}{Simulated Data, list of 3-d arrays}
  \item{\code{Points}}{Time points of stationarity transitions}
}
\description{
\code{MIC_sim} simulates multivariate time series for a sample of subjects, under stationary or piecewise stationary assumptions.
}
\examples{
\dontrun{
# Stationary simulation:
  x_stat <- MIC_sim(alpha = 0.9, nsub = 2, segs = 10, fs = 100)

# Sample adherence (alpha)
  sum(x_stat$C == x_stat$Ci[, 1]) / length(x_stat$C)

# Change of stationarity points
  x_stat$Points


# Non-stationary (piecewise stationary) simulation:

  x_nstat <- MIC_sim(alpha = 0.9, nsub = 2, segs = 10, fs = 100)
  x_nstat$Points
}
}
\references{
Qian Li, Damla Senturk, Catherine A. Sugar, Shanali Jeste, Charlotte DiStefano, Joel Frohlich, Donatello Telesca
  "\emph{Inferring Brain Signals Synchronicity from a Sample of EEG Readings}".
}
