% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC_helper.R
\name{pars}
\alias{pars}
\title{MIC_sim Helper: pars}
\usage{
pars(eta, M = 1.1, fs)
}
\arguments{
\item{eta}{numeric, peak location}

\item{M}{numeric, dispersion of power with narrower peak as M approximates 1, default at 1.1}

\item{fs}{sampling frequency}
}
\value{
vector of AR(2) coefficients
}
\description{
\code{pars} determines the AR(2) coefficient with desired oscillation properties: peak location, peak width and
 sampling frequency.
}
\seealso{
\code{\link{MIC_sim}} for its usage
}
