% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SpecOnly}
\alias{SpecOnly}
\title{Spectral densities averaged by overlapped sliding windows}
\usage{
SpecOnly(ts, lag, wn, win, overlap, specN)
}
\arguments{
\item{ts, }{3 dimensional array of time series}

\item{lag, }{integer, trunctation for spectral estimates in \code{\link{spec.parzen}}}

\item{wn, }{integer, maximal frequency as in \code{\link{specParzen}}}

\item{win, }{integer, moving average window size for spectral smoothing}

\item{overlap, }{integer, moving average overlap size for spectral smoothing}

\item{specN, }{integer, spectral resolution in \code{\link{specParzen}}}
}
\value{
3d array of spectral densities
}
\description{
\code{SpecOnly} calculates the normalized PSD averaged within sliding windows.
    For its usage, please refer to \code{\link{MIC_prep}}.
}
\seealso{
\code{\link{spec.parzen}} for spectral density estimates and \code{\link{MIC_prep}}
  for time series preprocessing before \code{\link{MIC}}
}
