% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{specParzen}
\alias{specParzen}
\title{Cpp function for Parzen window smoothed power spectral density}
\usage{
specParzen(ts, lag, maxf, outn)
}
\arguments{
\item{ts, }{vector of the time series}

\item{lag, }{integer of the max lag to truncate acf}

\item{maxf, }{integer of the max frequency to consider}

\item{outn, }{integer of the length of the spectral estimates}
}
\value{
vector of estimated spectral density of length \code{outn}
}
\description{
\code{specParzen} uses a parzen window smoothed autocovariance function
  to estimate the correlogram with trunction. Please refer to the R function
  \code{\link{spec.parzen}} for general usage.
}
