% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SpecSim}
\alias{SpecSim}
\title{Spectral similarity based on Total Variation Distance (TVD)}
\usage{
SpecSim(ts, lag, wn, win, overlap, specN)
}
\arguments{
\item{ts, }{3 dimensional array of time series}

\item{lag, }{integer, trunctation for spectral estimates in \code{\link{spec.parzen}}}

\item{wn, }{integer, maximal frequency as in \code{\link{specParzen}}}

\item{win, }{integer, moving average window size for spectral smoothing}

\item{overlap, }{integer, moving average overlap size for spectral smoothing}

\item{specN, }{integer, spectral resolution in \code{\link{specParzen}}}
}
\value{
3d array admissible to \code{\link{EigLap}}
}
\description{
\code{SpecSim} calculates the similarity between two normalized PSD, defined as
    their common area under the curves. It also correspond to \code{1-TVD(f,g)}
    for two normalized densities \code{f} and \code{g}. For its usage, please
    refer to \code{\link{MIC_prep}}.
}
\seealso{
\code{\link{spec.parzen}} for spectral density estimates and \code{\link{MIC_prep}}
  for time series preprocessing before \code{\link{MIC}}
}
